+++ src/packages/client/store/index.ts
@@ -1209,6 +1209,7 @@ class Store
   getTableSchema(...args: Parameters<DatabaseActions['getTableSchema']>) { return this.databaseActions.getTableSchema(...args); }
   setTableSchema(...args: Parameters<DatabaseActions['setTableSchema']>) { return this.databaseActions.setTableSchema(...args); }
   executeQuery(...args: Parameters<DatabaseActions['executeQuery']>) { return this.databaseActions.executeQuery(...args); }
+  executeQueryAndWait(...args: Parameters<DatabaseActions['executeQueryAndWait']>) { return this.databaseActions.executeQueryAndWait(...args); }
   executeSilentQuery(...args: Parameters<DatabaseActions['executeSilentQuery']>) { return this.databaseActions.executeSilentQuery(...args); }
   setSilentQueryResult(...args: Parameters<DatabaseActions['setSilentQueryResult']>) { return this.databaseActions.setSilentQueryResult(...args); }
   setQueryResult(...args: Parameters<DatabaseActions['setQueryResult']>) { return this.databaseActions.setQueryResult(...args); }
@@ -1248,7 +1249,7 @@ declare global {
 }
 
 // Increment this when Store class has breaking changes that require fresh instance
-const STORE_VERSION = 2;
+const STORE_VERSION = 3;
 
 // Singleton store instance - persisted on window for HMR
 function getOrCreateStore(): Store {